+++ chevah/utils/tests/normal/test_configuration_file.py
@@ -261,17 +261,6 @@ class TestFileConfigurationProxy(UtilsTestCase):
         self.assertEqual(u'1002', context.exception.event_id)
         config_file.close()
 
-    def test_getBoolean_bad_values(self):
-        '''Test bad value for remote_admin_enabled.'''
-        config_file = StringIO(
-            '[section]\n'
-            'bool_option = 3234\n')
-        config = FileConfigurationProxy(configuration_file=config_file)
-        config.load()
-        with self.assertRaises(UtilsError) as context:
-            config.getBoolean('section', 'bool_option')
-        self.assertEqual(u'1000', context.exception.event_id)
-
     def test_getString(self):
         """
         Check getString.
@@ -592,6 +581,198 @@ class TestFileConfigurationProxy(UtilsTestCase):
         self.assertEqual(
             7,
             config.getIntegerOrNone(u'some_section', u'some_int'))
+
+    def test_setInteger_int(self):
+        """
+        Set integer can be used for directly setting an integer number.
+        """
+        content = (
+            u'[some_section]\n'
+            u'some_int: 7\n'
+            )
+        config = self._getConfig(content=content)
+
+        config.setInteger('some_section', u'some_int', 10)
+
+        self.assertEqual(
+            10, config.getInteger(u'some_section', u'some_int'))
+
+    def test_setInteger_bad_value(self):
+        """
+        When setting an integer to a bad value, an error is raised and
+        previous value is kept.
+        """
+        content = (
+            u'[some_section]\n'
+            u'some_int: 7\n'
+            )
+        config = self._getConfig(content=content)
+
+        with self.assertRaises(UtilsError) as context:
+            config.setInteger('some_section', u'some_int', 'bad-value')
+
+        self.assertExceptionID(u'1001', context.exception)
+        self.assertContains('integer value', context.exception.message)
+        self.assertEqual(
+            7, config.getInteger(u'some_section', u'some_int'))
+
+    def test_setInteger_float(self):
+        """
+        Float values are floor rounded to integers.
+        """
+        content = (
+            u'[some_section]\n'
+            u'some_int: 7\n'
+            )
+        config = self._getConfig(content=content)
+
+        config.setInteger('some_section', u'some_int', 100.6)
+
+        self.assertEqual(
+            100, config.getInteger(u'some_section', u'some_int'))
+
+    def test_getBoolean_valid(self):
+        """
+        A boolean value is read if stored in human readable boolean.
+        """
+        content = (
+            '[section]\n'
+            'bool_option: YeS\n')
+        config = self._getConfig(content=content)
+
+        self.assertTrue(config.getBoolean('section', 'bool_option'))
+
+    def test_getBoolean_invalid(self):
+        """
+        An error is raised when trying to rad a bad boolean value.
+        """
+        content = (
+            '[section]\n'
+            'bool_option = 3234\n')
+        config = self._getConfig(content=content)
+
+        with self.assertRaises(UtilsError) as context:
+            config.getBoolean('section', 'bool_option')
+
+        self.assertEqual(u'1000', context.exception.event_id)
+
+    def test_setBoolean_valid(self):
+        """
+        It can set a boolean value specified as free form.
+        """
+        content = (
+            u'[some_section]\n'
+            u'some_boolean: yes\n'
+            )
+        config = self._getConfig(content=content)
+
+        config.setBoolean('some_section', u'some_boolean', 'No')
+
+        self.assertFalse(config.getBoolean(u'some_section', u'some_boolean'))
+
+    def test_setBoolean_invalid(self):
+        """
+        When a bad boolean is set, the old value is kept and an error is
+        raised.
+        """
+        content = (
+            u'[some_section]\n'
+            u'some_boolean: yes\n'
+            )
+        config = self._getConfig(content=content)
+
+        with self.assertRaises(UtilsError) as context:
+            config.setBoolean('some_section', u'some_boolean', 'bad-value')
+
+        self.assertExceptionID(u'1001', context.exception)
+        self.assertContains('boolean value', context.exception.message)
+        self.assertTrue(config.getBoolean(u'some_section', u'some_boolean'))
+
+    def test_booleanConverter_valid(self):
+        """
+        Convert free from boolean into standard pyton boolean values.
+        """
+        config = self._getConfig(content='')
+
+        for value in [True, 1, '1', 'tRuE', 'yEs']:
+            self.assertTrue(config._booleanConverter(value))
+
+        for value in [False, 0, '0', 'falSE', 'nO']:
+            self.assertFalse(config._booleanConverter(value))
+
+    def test_booleanConverter_invalid(self):
+        """
+        An error is raised when value can not be converted.
+        """
+        config = self._getConfig(content='')
+
+        with self.assertRaises(ValueError):
+            config._booleanConverter('no-boolean')
+
+        with self.assertRaises(ValueError):
+            config._booleanConverter(object())
+
+        with self.assertRaises(ValueError):
+            config._booleanConverter(4)
+
+    def test_getFload_valid(self):
+        """
+        A float value is can be read.
+        """
+        content = (
+            '[section]\n'
+            'float_option: 1.43\n')
+        config = self._getConfig(content=content)
+
+        self.assertEqual(1.43, config.getFloat('section', 'float_option'))
+
+    def test_getFload_invalid(self):
+        """
+        An error is raised when value is not float.
+        """
+        content = (
+            '[section]\n'
+            'float_option: bla\n')
+        config = self._getConfig(content=content)
+
+        with self.assertRaises(UtilsError) as context:
+            config.getFloat('section', 'float_option')
+
+        self.assertExceptionID(u'1000', context.exception)
+        self.assertContains(
+            'floating number value', context.exception.message)
+
+    def test_setFload_valid(self):
+        """
+        A float value is can be set as float or string.
+        """
+        content = (
+            '[section]\n'
+            'float_option: 0\n')
+        config = self._getConfig(content=content)
+
+        config.setFloat('section', 'float_option', 2.45)
+        self.assertEqual(2.45, config.getFloat('section', 'float_option'))
+
+        config.setFloat('section', 'float_option', u'3.45')
+        self.assertEqual(3.45, config.getFloat('section', 'float_option'))
+
+    def test_setFload_invalid(self):
+        """
+        When setting an invalid float value, error is raised and previous
+        value is kept.
+        """
+        content = (
+            '[section]\n'
+            'float_option: 2.34\n')
+        config = self._getConfig(content=content)
+
+        with self.assertRaises(UtilsError) as context:
+            config.setFloat('section', 'float_option', 'bad-value')
+
+        self.assertExceptionID(u'1001', context.exception)
+        self.assertContains(
+            'floating number value', context.exception.message)
 
 
 class DummyConfigurationFileMixin(ConfigurationFileMixin):